refactor: extract database bindings
